templates/backends/python/requirements.txt
flask==3.1.0
flask-cors==5.0.1
mysql-connector-python==9.2.0
oracledb==3.1.1
